chore: setup router and import aliases
