+++ .eslintrc.json
@@ -7,21 +7,47 @@
     },
     "rules": {
         "strict": 0,
-        "quotes": [2, "single"],
+        "quotes": [
+            2,
+            "single"
+        ],
         "no-else-return": 0,
         "new-cap": 0,
         "no-console": 0,
-        "import/no-unresolved": [2, { "caseSensitive": false } ],
-        "no-unused-vars": ["error", { "vars": "all", "args": "none" }],
+        "import/no-unresolved": [
+            2,
+            {
+                "caseSensitive": false
+            }
+        ],
+        "no-unused-vars": [
+            "error",
+            {
+                "vars": "all",
+                "args": "none"
+            }
+        ],
         "no-underscore-dangle": 0,
         "arrow-body-style": 0,
-        "one-var": ["error", { "uninitialized": "always", "initialized": "never" }],
-        "one-var-declaration-per-line": ["error", "initializations"],
-        "max-len": ["error", 200],
+        "one-var": [
+            "error",
+            {
+                "uninitialized": "always",
+                "initialized": "never"
+            }
+        ],
+        "one-var-declaration-per-line": [
+            "error",
+            "initializations"
+        ],
+        "max-len": [
+            "error",
+            200
+        ],
         "no-extra-parens": 0,
         "no-restricted-syntax": [
-          0,
-          "DebuggerStatement"
+            0,
+            "DebuggerStatement"
         ],
         "no-debugger": "warn",
         "react/jsx-uses-react": 2,
@@ -33,7 +59,7 @@
         "react/jsx-filename-extension": 0,
         "jsx-a11y/click-events-have-key-events": 0,
         "jsx-a11y/no-noninteractive-element-interactions": 0,
-        "react/jsx-one-expression-per-line":0,
+        "react/jsx-one-expression-per-line": 0,
         "no-useless-escape": 0,
         "jsx-a11y/no-static-element-interactions": 0
     },
@@ -44,6 +70,11 @@
         "ecmaFeatures": {
             "jsx": true,
             "sourceType": "module"
+        }
+    },
+    "settings": {
+        "import/resolver": {
+            "babel-module": {}
         }
     }
 }
