--- babel.config.json
@@ -1,13 +1,0 @@
-{
-  "plugins": [
-    "@babel/plugin-transform-runtime",
-    "@babel/plugin-proposal-class-properties",
-    "@babel/plugin-proposal-object-rest-spread"
-  ],
-  "presets": [
-    ["@babel/react"],
-    ["@babel/preset-env", {
-      "targets": "> 0.25%, not dead"
-    }]
-  ]
-}
