+++ jsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "baseUrl": ".",
+        "paths": {
+            "#*": [
+                "src/*"
+            ]
+        }
+    },
+    "exclude": ["node_modules"]
+}
