+++ package.json
@@ -19,11 +19,13 @@
     "@babel/preset-react": "^7.17.12",
     "autoprefixer": "^10.4.7",
     "babel-loader": "^8.2.5",
+    "babel-plugin-module-resolver": "^4.1.0",
     "css-loader": "^6.7.1",
     "dotenv-safe": "^8.2.0",
     "dotenv-webpack": "^7.1.0",
     "eslint": "^8.17.0",
     "eslint-config-airbnb": "^19.0.4",
+    "eslint-import-resolver-babel-module": "^5.3.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.30.0",
@@ -47,6 +49,7 @@
     "lodash.debounce": "^4.0.8",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-swipeable": "^7.0.0"
   }
 }
