+++ src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 
-import App from './app';
 import './style.scss';
+
+import App from './app';
 
 const root = createRoot(document.getElementById('main'));
 root.render(<App />);
